+++ hugolib/testhelpers_test.go
@@ -300,9 +300,6 @@ defaultContentLanguageInSubdir = true
 [permalinks]
 other = "/somewhere/else/:filename"
 
-[blackfriday]
-angledQuotes = true
-
 [Taxonomies]
 tag = "tags"
 
@@ -311,8 +308,6 @@ tag = "tags"
 weight = 10
 title = "In English"
 languageName = "English"
-[Languages.en.blackfriday]
-angledQuotes = false
 [[Languages.en.menu.main]]
 url    = "/"
 name   = "Home"
