+++ markup/converter/converter.go
@@ -119,11 +119,10 @@ func (b Bytes) Bytes() []byte {
 
 // DocumentContext holds contextual information about the document to convert.
 type DocumentContext struct {
-	Document        any // May be nil. Usually a page.Page
-	DocumentID      string
-	DocumentName    string
-	Filename        string
-	ConfigOverrides map[string]any
+	Document     any // May be nil. Usually a page.Page
+	DocumentID   string
+	DocumentName string
+	Filename     string
 }
 
 // RenderContext holds contextual information about the content to render.
